+++ core/ui/build.gradle.kts
@@ -25,11 +25,11 @@ android {
         }
     }
     compileOptions {
-        sourceCompatibility = JavaVersion.VERSION_11
-        targetCompatibility = JavaVersion.VERSION_11
+        sourceCompatibility = JavaVersion.VERSION_17
+        targetCompatibility = JavaVersion.VERSION_17
     }
     kotlinOptions {
-        jvmTarget = "11"
+        jvmTarget = "17"
     }
 }
 
@@ -59,4 +59,7 @@ dependencies {
 
     // icon
     implementation(libs.androidx.material.icons.extended)
+
+    // datetime
+    implementation(libs.kotlinx.datetime)
 }
